feat(sharing): [#FOR-433] separate sharing API points in a new dedicated controller
